refactor: inline string identifier rules
